Remove else after return

+++ runtime/lib/ttnn/include/tt/runtime/ttnn/utils.h
@@ -51,7 +51,7 @@ Tensor createRuntimeTensorFromTTNN(const ::ttnn::Tensor &tensor);
 // TODO: (#1435): Fix int types across shapes
 //
 inline std::vector<uint32_t>
-toShapeFromFBShape(const flatbuffers::Vector<int64_t> &vec) {
+toShapeFromFBShape(const flatbuffers::Vector<int32_t> &vec) {
   return std::vector<uint32_t>(vec.begin(), vec.end());
 }
